+++ test/parser.js
@@ -1,0 +1,50 @@
+var buster  = require('buster-node');
+var fs      = require('fs');
+var parser  = require('../lib/parser');
+var referee = require('referee');
+var yamljs  = require('yaml-js');
+var assert  = referee.assert;
+
+buster.testCase('parser - parse', {
+  setUp: function () {
+    this.mockFs     = this.mock(fs);
+    this.mockYamljs = this.mock(yamljs);
+
+    this.mockFs.expects('readFileSync').withExactArgs('somefile').returns('sometext');
+  },
+  'should add file info to test': function (done) {
+    var data = [{ command: 'whoami', description: 'somedesc' }];
+    this.mockYamljs.expects('load').withExactArgs('sometext').returns(data);
+    parser.parse('somefile', function (err, tests) {
+      assert.isNull(err);
+      assert.equals(tests.length, 1);
+      assert.equals(tests[0].file, 'somefile');
+      done();
+    });
+  },
+  'should merge parameters into command': function (done) {
+    var data = [
+      { params: { message: 'some message' }},
+      { command: 'echo "{message}"', description: 'somedesc' }
+    ];
+    this.mockYamljs.expects('load').withExactArgs('sometext').returns(data);
+    parser.parse('somefile', function (err, tests) {
+      assert.isNull(err);
+      assert.equals(tests.length, 1);
+      assert.equals(tests[0].command, 'echo "some message"');
+      done();
+    });
+  },
+  'should replace parameter with blank when there is no associated parameter': function (done) {
+    var data = [
+      { command: 'echo "{message}"', description: 'somedesc' }
+    ];
+    this.mockYamljs.expects('load').withExactArgs('sometext').returns(data);
+    parser.parse('somefile', function (err, tests) {
+      assert.isNull(err);
+      assert.equals(tests.length, 1);
+      assert.equals(tests[0].command, 'echo ""');
+      done();
+    });
+  }
+});