+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.6-pre
-*
+* Add parameter support for test commands
 
 ### 0.0.5
 * Fix sample test file on README.